Testing CSV display using panda and pyscript #2
