+++ HTML templates/css testing/csv_files/DecodedOutput.txt
@@ -1,0 +1,29 @@
+PRL1 will be taught by Prof_1 on M/W/(F) at 8:00 blocked off in column 1
+OPR1 will be taught by Prof_1 on M/W/(F) at 9:30 blocked off in column 1
+ORG1 will be taught by Prof_5 on M/W/F at 11:00 blocked off in column 1
+ARI1 will be taught by Prof_5 on M/W/(F) at 2:00 blocked off in column 1
+EXP1 will be taught by Prof_1 on M/W/(F) at 3:30 blocked off in column 1
+ALG1 will be taught by Prof_1 on T/TH at 8:30 blocked off in column 1
+TOC1 will be taught by Prof_2 on T/TH at 10:00 blocked off in column 1
+DAT1 will be taught by Prof_5 on T/TH at 11:30 blocked off in column 1
+NUA1 will be taught by Prof_4 on T/TH at 1:00 blocked off in column 1
+DIS1 will be taught by Prof_3 on T/TH at 2:30 blocked off in column 1
+CS11 will be taught by Prof_2 on M/W/F at 9:30 blocked off in column 2
+CAP1 will be taught by Prof_1 on M/W/F at 11:00 blocked off in column 2
+CS21 will be taught by Prof_5 on M/W/F at 2:00 blocked off in column 2
+SYS1 will be taught by Prof_5 on T/TH at 2:30 blocked off in column 2
+PRL1 will be taught by Prof_1 on M/W/(F) at 8:00 blocked off in column 1
+OPR1 will be taught by Prof_1 on M/W/(F) at 9:30 blocked off in column 1
+ORG1 will be taught by Prof_5 on M/W/F at 11:00 blocked off in column 1
+ARI1 will be taught by Prof_5 on M/W/(F) at 2:00 blocked off in column 1
+EXP1 will be taught by Prof_1 on M/W/(F) at 3:30 blocked off in column 1
+ALG1 will be taught by Prof_1 on T/TH at 8:30 blocked off in column 1
+TOC1 will be taught by Prof_2 on T/TH at 10:00 blocked off in column 1
+DAT1 will be taught by Prof_5 on T/TH at 11:30 blocked off in column 1
+NUA1 will be taught by Prof_4 on T/TH at 1:00 blocked off in column 1
+DIS1 will be taught by Prof_3 on T/TH at 2:30 blocked off in column 1
+CS11 will be taught by Prof_2 on M/W/F at 9:30 blocked off in column 2
+CAP1 will be taught by Prof_1 on M/W/F at 11:00 blocked off in column 2
+CS21 will be taught by Prof_5 on M/W/F at 2:00 blocked off in column 2
+SYS1 will be taught by Prof_5 on T/TH at 2:30 blocked off in column 2
+Success! 14 courses offered and printed!
