+++ HTML templates/css testing/csv_files/MathTemplateCourseRestrictions.csv
@@ -1,0 +1,15 @@
+STA1,FALSE,m330,$
+MC11,TRUE,m330,$
+MC21,TRUE,m330,$
+EPM1,FALSE,$
+MC31,TRUE,LNA1,AST1,MCP1,DAN1,PMS1,DEQ1,RAN1,GEO1,CMA1,SPT1,m330,$
+LNA1,TRUE,MC31,AST1,MCP1,DAN1,PMS1,DEQ1,RAN1,GEO1,CMA1,SPT1,m330,$
+AST1,FALSE,MC31,LNA1,MCP1,AST1,GEO1,CMA1,SPT1,m330,$
+MCP1,TRUE,MC31,LNA1,AST1,DAN1,GEO1,CMA1,SPT1,m330,$
+DAN1,FALSE,MC31,LNA1,AST1,MCP1,GEO1,CMA1,SPT1,m330,$
+GEO1,FALSE,CMA1,SPT1,MC31,LNA1,AST1,MCP1,DAN1,PMS1,DEQ1,RAN1,m330,$
+CMA1,FALSE,GEO1,SPT1,MC31,LNA1,AST1,MCP1,DAN1,PMS1,DEQ1,RAN1,m330,$
+SPT1,FALSE,GEO1,CMA1,MC31,LNA1,AST1,MCP1,DAN1,PMS1,DEQ1,RAN1,m330,$
+RAN1,FALSE,MC31,LNA1,GEO1,CMA1,SPT1,PMS1,DEQ1,m330,$
+PMS1,FALSE,MC31,LNA1,GEO1,CMA1,SPT1,RAN1,DEQ1,m330,$
+DEQ1,FALSE,MC31,LNA1,GEO1,CMA1,SPT1,RAN1,PMS1,m330,$
