+++ HTML templates/css testing/csv_files/MathTemplateFacultyRestrictions.csv
@@ -1,0 +1,9 @@
+Prof_1,FALSE,STA1,GEO1,$
+Prof_2,FALSE,MC11,CMA1,$
+Prof_3,FALSE,EPM1,SPT1,$
+Prof_4,FALSE,LNA1,PMS1,$
+Prof_5,FALSE,AST1,RAN1,$
+Prof_6,FALSE,MC21,DEQ1,$
+Prof_7,FALSE,MC31,$
+Prof_8,FALSE,MCP1,$
+Prof_9,FALSE,DAN1,$
