+++ HTML templates/css testing/csv_files/WritingCSVFiles.md
@@ -1,0 +1,18 @@
+# Formatting for CSV Files
+While there are 4 main files to deal with when generating Linear Programs, only 2 need to be created. This file is made to outline the format of creating csv files so you can create your own Linear Programs.
+
+## Course Restrictions
+
+The course restrictions are the first of two files you should create, as it outlines all of the courses and sections offered. Each course/ section name should be abbreviated to four characters, 3 for the course and one for the section. For example, Section 1 of Discrete Math could be written as DIS1. These abbreviations must be 4 characters long and must be consistent throughout the course and faculty restrictions to ensure the code knows which course is which. Other than being consistent, any abbreviation can be chosen for easy understanding for the lp creator.
+
+Next, each course will be at the left most column of the csv file. The next thing to add is a true or false in all caps to differentiate whether or not the course must meet for 4 hours a week. When true, this forces courses to only be available on M/W/F slots, since T/TH slots aren't possible. Then, after the true or false list all of the other courses or times that this particular course cannot be at the same time as. When all restraints are listed, add a $ to indicate the end of the row for the code. If you're having trouble, you can follow the example in math or cs templates. Here is a link to the Cs template to see what it could look like https://github.com/SUDrAnthonyCourses/ScheduleOptimizer/blob/Muell/csv_files/csTemplateCourseRestrictions.csv. NOTE: all courses must have a true/ false and a $ for the code to work, even if there are no constraints for a given course.
+## Faculty Restrictions
+The faculty constraints are where the courses are assigned to the professors and the other file you'll have to create to make your own LP. In the right most spot in every row should be the professor's name, this has no character limit, unlike the courses. Next, there is a True/False to indicate whether or not a professor must teach outside of the prime time hours. If this is True, then at least one course will be scheduled outside of the prime time hours. Then, all of the courses listed are what a given professor is assigned to, followed by any time slots that a professor can't teach at. The final slot must also be a $ to indicate the end of a row. Again, if you're having trouble you can consult the math or CS templates as a reference, here is the link to the CS Template: https://github.com/SUDrAnthonyCourses/ScheduleOptimizer/blob/Muell/csv_files/csTemplateFacultyRestrictions.csv. Make sure that all courses are assigned to one professor, otherwise the readoutput file may look a little off
+
+NOTE: when creating a LP keep the number of courses below 30 and the total number of constraints below 220 (A constraint is any conflicting course or time or TRUE value in either the course or faculty restrictions). Going above this can cause the glpsol to have abnormally high run-times of over an hour, and this spike comes seemingly out of nowhere. This is still being investigated and these numbers are only from tests on my machine, they could be higher or lower depending on your computer.
+### LP File
+
+After plugging the course and faculty restrictions into GenerateLp.py, an output file with the file type .lp (or .txt) will be created. This will be what is plugged into glpsol, which will then run and create a solution file with a .txt file type.
+### Solution File
+After running it in glpsol, you'll have this solution file which can then be plugged into ReadOutput.py. This will then print (or write to a file) the assignments generated by glpsol. These are organized by column number, then meeting time, then course name for easy schedule building. If you choose to have it written to a file, it will contain the exact same information as what is printed out.
+
