+++ HTML templates/css testing/csv_files/csTemplateCourseRestrictions.csv
@@ -1,0 +1,14 @@
+CS11,TRUE,CS21,EXP1,ORG1,ALG1,m330,$
+CS21,TRUE,CS11,EXP1,ORG1,ALG1,m330,$
+DIS1,FALSE,ORG1,PRL1,ARI1,TOC1,DAT1,OPR1,NUA1,m330,$
+ORG1,TRUE,DIS,PRL1,ARI1,TOC1,DAT1,OPR1,NUA1,CS11,CS21,EXP1,ALG1,m330,$
+PRL1,FALSE,DIS,ORG1,ARI1,TOC1,DAT1,OPR1,NUA1,m330,$
+ALG1,FALSE,SYS1,CAP1,ARI1,TOC1,DAT1,OPR1,NUA1,CS11,CS21,EXP1,ORG1,m330,$
+SYS1,FALSE,ALG1,CAP1,ARI1,TOC1,DAT1,OPR1,NUA1,m330,$
+CAP1,TRUE,ALG1,SYS1,ARI1,TOC1,DAT1,OPR1,NUA1,m330,$
+ARI1,FALSE,TOC1,DAT1,OPR1,NUA1,DIS1,ORG1,PRL1,ALG1,SYS1,CAP1,m330,$
+TOC1,FALSE,ARI1,DAT1,OPR1,NUA1,DIS1,ORG1,PRL1,ALG1,SYS1,CAP1,m330,$
+DAT1,FALSE,ARI1,TOC1,OPR1,NUA1,DIS1,ORG1,PRL1,ALG1,SYS1,CAP1,m330,$
+OPR1,FALSE,ARI1,TOC1,DAT1,NUA1,DIS1,ORG1,PRL1,ALG1,SYS1,CAP1,m330,$
+NUA1,FALSE,ARI1,TOC1,DAT1,OPR1,DIS1,ORG1,PRL1,ALG1,SYS1,CAP1,m330,$
+EXP1,FALSE,CS11,CS21,ORG1,ALG1,$
