+++ HTML templates/css testing/csv_files/csTemplateFacultyRestrictions.csv
@@ -1,0 +1,5 @@
+Prof_1,TRUE,ALG1,CAP1,PRL1,OPR1,EXP1,m200,$
+Prof_2,FALSE,CS11,TOC1,$
+Prof_3,FALSE,DIS1,$
+Prof_4,FALSE,NUA1,$
+Prof_5,FALSE,CS21,ORG1,DAT1,SYS1,ARI1,$
