+++ HTML templates/css testing/dummy.csv
@@ -1,0 +1,32 @@
+Numeric,Numeric-2,Numeric-Suffix
+1,01,1st
+2,02,2nd
+3,03,3rd
+4,04,4th
+5,05,5th
+6,06,6th
+7,07,7th
+8,08,8th
+9,09,9th
+10,10,10th
+11,11,11th
+12,12,12th
+13,13,13th
+14,14,14th
+15,15,15th
+16,16,16th
+17,17,17th
+18,18,18th
+19,19,19th
+20,20,20th
+21,21,21st
+22,22,22nd
+23,23,23rd
+24,24,24th
+25,25,25th
+26,26,26th
+27,27,27th
+28,28,28th
+29,29,29th
+30,30,30th
+31,31,31st
